chore(popover): split demo
