+++ src/packages/__VUE/popover/demo/index.vue
@@ -20,7 +20,7 @@
   </Demo>
 </template>
 <script setup lang="ts">
-import { useTranslate } from '../../../utils';
+import { useTranslate } from '@/sites/utils';
 import Basic from './basic.vue';
 import Option from './option.vue';
 import Content from './content.vue';
